reorganize testfile crate
